Move ThemeMonitor to theming package
